Update function calldata array builder script, rename to backend/safeTxArrayBuilder.ts
